% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_on_known.R
\name{condition_on_known}
\alias{condition_on_known}
\title{Condition on known cases}
\usage{
condition_on_known(
  sims,
  days_since_end_seed = NULL,
  lower_bound = NULL,
  upper_bound = NULL
)
}
\arguments{
\item{sims}{A dataframe as produced by \code{scenario_analysis}.}

\item{days_since_end_seed}{Numeric, the number of days since the end of the seeding event.}

\item{lower_bound}{Numeric, the estimated lower bound on cases.}

\item{upper_bound}{Numeric, the estimated upper bound on cases.}
}
\value{

}
\description{
Condition on known cases
}
\examples{


}
