% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_delay_sample_func.R
\docType{data}
\name{fitted_delay_sample_func}
\alias{fitted_delay_sample_func}
\title{Fitted delay sampling function}
\format{A function with a single argument n that accepts integer values. Returns a vector of sampled reporting delays.}
\usage{
fitted_delay_sample_func
}
\description{
A function fitted to linelist data to estimate reporting data in model simultions. Update using \code{data-raw/fitted_delay_sample_func.R}.
}
\keyword{datasets}
