% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linelist.r
\name{get_linelist}
\alias{get_linelist}
\title{Get the latest compiled linelist}
\usage{
get_linelist()
}
\value{
a tibble with the lineist
delays to randomly sample
}
\description{
Get the latest compiled linelist
}
\examples{

## Example
\dontrun{
get_linelist()
}

## Code
get_linelist
}
\author{
Sebastian Funk \href{mailto:sebastian.funk@lshtm.ac.uk}{sebastian.funk@lshtm.ac.uk}
}
