% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rep_sample_fun.r
\name{get_rep_sample_fun}
\alias{get_rep_sample_fun}
\title{Get a sampling function for reporting delays}
\usage{
get_rep_sample_fun()
}
\value{
a function that takes one parameter, \code{n}, the number of reporting
delays to randomly sample
}
\description{
This connects to the Google Sheet containing the collated linelist; the
first time this is run, it may require authentiction with Google
}
\examples{

## Example
\dontrun{
get_rep_sample_fun()
}

## Code 
get_rep_sample_fun
}
\author{
Sebastian Funk \href{mailto:sebastian.funk@lshtm.ac.uk}{sebastian.funk@lshtm.ac.uk}
}
