% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_duration_size_table.R
\name{make_duration_size_table}
\alias{make_duration_size_table}
\title{Make seeding event duration vs event size table}
\usage{
make_duration_size_table(results = NULL)
}
\arguments{
\item{results}{The output from \code{summarised_end_r0}}
}
\value{
A data.frame comparing transmission event size with duration
}
\description{
Make seeding event duration vs event size table
}
\examples{

## Code
make_duration_size_table
}
\author{
Sam Abbott
}
