% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_per_accepted_tile.R
\name{plot_per_accepted_tile}
\alias{plot_per_accepted_tile}
\title{Plot a heat map of accepted samples by scenario}
\usage{
plot_per_accepted_tile(proportion_allowed = NULL)
}
\arguments{
\item{proportion_allowed}{data.frame of allowed proportions as
produced by \code{proportion_allowed_by_condition}
by \code{condition_and_report_on_cases}.}
}
\value{
A ggplot2 heat-map of accepted percentages by scenario
}
\description{
Plot a heat map of accepted samples by scenario
}
\examples{

## Code
plot_per_accepted_tile
}
\author{
Sam Abbott
}
