% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_R0_density.R
\name{plot_R0_density}
\alias{plot_R0_density}
\title{Plot density plots of the basic reproduction number}
\usage{
plot_R0_density(sims = NULL)
}
\arguments{
\item{sims}{A data.table from \code{scenario_analysis}.}
}
\value{
A ggplot2 plot of R0 density over the scenarios explored
}
\description{
Plot density plots of the basic reproduction number
}
\examples{

## Code
plot_R0_density
}
