% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scenario_tile.R
\name{plot_scenario_tile}
\alias{plot_scenario_tile}
\title{Plot a heat map of accepted samples by scenario}
\usage{
plot_scenario_tile(conditioned_output = NULL)
}
\arguments{
\item{conditioned_output}{A list of dataframes as produced
by \code{condition_and_report_on_cases}.}
}
\value{

}
\description{
Plot a heat map of accepted samples by scenario
}
\examples{


}
\author{
Sam Abbott
}
