% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_allowed_by_condition.R
\name{proportion_allowed_by_condition}
\alias{proportion_allowed_by_condition}
\title{Proportion of samples per scenario allowed by condition}
\usage{
proportion_allowed_by_condition(
  sims,
  samples = NULL,
  group_var = c("scenario", "event_duration", "event_size", "serial_mean", "upper_R0",
    "lower_R0")
)
}
\arguments{
\item{sims}{A data.table from \code{scenario_analysis}.}

\item{samples}{Numeric, the number of samples that were used.}

\item{group_var}{Character vector, the grouping variables to use to split scenarios.}
}
\value{
A data.table listing the proportion allowed by scenario
}
\description{
Proportion of samples per scenario allowed by condition
}
\examples{


sims <- data.table::data.table(sample = 1:10, event_duration = 1,
                               event_size = 1:10)
proportion_allowed_by_condition(sims, samples = 100, 
                                group_var = c("event_duration"))
}
\author{
Sam Abbott
}
