% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_by_condition.R
\name{restrict_by_condition}
\alias{restrict_by_condition}
\title{Restrict scenarios based on allowed scenarios}
\usage{
restrict_by_condition(sims, allowed_scenarios)
}
\arguments{
\item{sims}{A dataframe as produced by \code{scenario_analysis}.}

\item{allowed_scenarios}{A dataframe of allowed scenarios. Must include a scenario and sample
variable.}
}
\value{

}
\description{
Restrict scenarios based on allowed scenarios
}
\examples{


}
