% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_analysis.R
\name{scenario_analysis}
\alias{scenario_analysis}
\title{Run a scenario analyis}
\usage{
scenario_analysis(
  scenarios = NULL,
  sampled_and_set_parameters = NULL,
  delay_sample_func = NULL,
  show_progress = FALSE,
  kept_times
)
}
\arguments{
\item{scenarios}{Dataframe of potential scenarios}

\item{sampled_and_set_parameters}{Dataframe of sampled and fixed parameter values}

\item{delay_sample_func}{Function to generate sample reporting delays}

\item{show_progress}{Logical, defaults to FALSE. Show progress be shown.}

\item{kept_times}{Numeric, a vector of timepoints to keep information on. Defaults to all
time points}
}
\value{
A nested dataframe of scenarios combined with model simulations
}
\description{
Run a scenario analyis
}
\examples{

## Code 
scenario_analysis
}
\author{
Sam Abbott
}
