% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_end_r0.R
\name{summarise_end_r0}
\alias{summarise_end_r0}
\title{Summarise R0 estimates across samples}
\usage{
summarise_end_r0(sims, group_var = c("event_size", "event_duration"))
}
\arguments{
\item{sims}{A data.table from \code{scenario_analysis}.}

\item{group_var}{Character vector indicating the groups to show. Defaults to
\code{event_duration} and \code{event_size}}
}
\value{
A dataframe of R0 estimates for each scenario
}
\description{
Summarise R0 estimates across samples
}
\examples{

## Code
summarise_end_r0
}
\author{
Sam Abbott
}
