% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WuhanSeedingVsTransmission-package.R
\docType{package}
\name{WuhanSeedingVsTransmission-package}
\alias{WuhanSeedingVsTransmission}
\alias{WuhanSeedingVsTransmission-package}
\title{WuhanSeedingVsTransmission: Evaluate seeding event size vs duration compared to transmission in the Wuhan outbreak}
\description{
Evaluate seeding event size vs duration compared to transmission in the Wuhan outbreak.
}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

Authors:
\itemize{
  \item Joel Hellewell \email{joel.hellewell@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-2683-0849}{ORCID})
  \item James Munday \email{james.munday@lshtm.ac.uk}
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}

}
\keyword{internal}
