% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_and_report_on_cases.R
\name{condition_and_report_on_cases}
\alias{condition_and_report_on_cases}
\title{Condition on cases and return accepted sample size by scenario}
\usage{
condition_and_report_on_cases(
  sims,
  condition_date = NULL,
  lower_bound = NULL,
  upper_bound = NULL,
  samples = NULL,
  end_of_seed_date = "2019-12-31"
)
}
\arguments{
\item{sims}{A dataframe as produced by \code{scenario_analysis}.}

\item{condition_date}{A character string in the following format \code{"2020-01-01"}. Sets the
date on which to condition the data.}

\item{lower_bound}{Numeric, the estimated lower bound on cases.}

\item{upper_bound}{Numeric, the estimated upper bound on cases.}

\item{samples}{Numeric, the number of samples that were used.}
}
\value{

}
\description{
Condition on cases and return accepted sample size by scenario
}
\examples{


}
