% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_on_known.R
\name{condition_on_known}
\alias{condition_on_known}
\title{Condition on known cases}
\usage{
condition_on_known(
  sims,
  days_since_end_seed = NULL,
  lower_bound = NULL,
  upper_bound = NULL
)
}
\arguments{
\item{sims}{A data.table from \code{scenario_analysis}.}

\item{days_since_end_seed}{Numeric, the number of days since the end of the seeding event.}

\item{lower_bound}{Numeric, the estimated lower bound on cases.}

\item{upper_bound}{Numeric, the estimated upper bound on cases.}
}
\value{
A data.table of conditioned scenarios
}
\description{
Condition on known cases
}
\examples{

sims <- data.table::data.table(
                   time = rep(1:10, 10),
                   size = rep(1:10, 10),
                   sample = unlist(lapply(1:10, function(.) {rep(., 10)})),
                   scenario = c(rep(1, 5), rep(1, 5)),
                   tmp = c(rep(1, 5), rep(1, 5)),
                   event_duration = 1
                  )

 sims <- data.table::data.table(scenario = 1:10, sample = 1:10, size = 1:10,
                                event_duration = rep(c(1, 2), 5), time = 1:10)
                                
condition_on_known(sims, upper_bound = 5, lower_bound = 3, days_since_end_seed = 0)
}
\author{
Sam Abbott
}
