% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scenario_tile.R
\name{plot_scenario_tile}
\alias{plot_scenario_tile}
\title{Plot a heat map of accepted samples by scenario}
\usage{
plot_scenario_tile(proportion_allowed = NULL)
}
\arguments{
\item{proportion_allowed}{data.frame of allowed proportions as
produced by \code{proportion_allowed_by_condition}
by \code{condition_and_report_on_cases}.}
}
\value{

}
\description{
Plot a heat map of accepted samples by scenario
}
\examples{


}
\author{
Sam Abbott
}
