% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_allowed_by_condition.R
\name{proportion_allowed_by_condition}
\alias{proportion_allowed_by_condition}
\title{Proportion of samples per scenario allowed by condition}
\usage{
proportion_allowed_by_condition(sims, samples = NULL)
}
\arguments{
\item{sims}{A dataframe as produced by \code{scenario_analysis}.}

\item{samples}{Numeric, the number of samples that were used.}
}
\value{

}
\description{
Proportion of samples per scenario allowed by condition
}
\examples{


}
\author{
Sam Abbott
}
