% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_by_condition.R
\name{restrict_by_condition}
\alias{restrict_by_condition}
\title{Restrict scenarios based on allowed scenarios}
\usage{
restrict_by_condition(
  sims,
  allowed_scenarios,
  restriction_var = c("scenario", "sample")
)
}
\arguments{
\item{sims}{A data.table from \code{scenario_analysis}.}

\item{allowed_scenarios}{A data.table of allowed scenarios. Must include a scenario and sample
variable.}

\item{restriction_var}{Character vector of variables to use to restrict scenarios. Defaults to "scenario" and
"sample".}
}
\value{
A data frame of scenarios restricted by the allowed scenarios
}
\description{
Restrict scenarios based on allowed scenarios
}
\examples{

sims <- data.table::data.table(scenario = 1:10, sample = 1:10, value = 1:10)

allowed_scenarios <- data.table::data.table(scenario = c(1,4,6), sample = c(1, 4, 6))

restrict_by_condition(sims, allowed_scenarios)

}
\author{
Sam Abbott
}
