% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario_grid.R
\name{run_scenario_grid}
\alias{run_scenario_grid}
\title{Run the scenario analysis grid.}
\usage{
run_scenario_grid(
  end_date = NULL,
  samples = 1,
  upper_case_bound = NULL,
  show_progress = FALSE
)
}
\arguments{
\item{end_date}{Character string in the format \code{"2020-01-01"}. The date to run the model too.}

\item{samples}{Numeric, defaults to 1. The number of samples to take for each scenario.}

\item{upper_case_bound}{Numeric, defaults to \code{NULL}. The upper bound on the number of cases that will be
modelled}

\item{show_progress}{Logical, defaults to FALSE. Show progress be shown.}
}
\value{

}
\description{
Run the scenario analysis grid.
}
\examples{


grid_results <- run_scenario_grid("2020-01-25", samples = 1, 
                  upper_case_bound = 100, show_progress = TRUE)
                  
grid_results
}
\author{
Sam Abbott
}
