% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Negative binomial branching process model}
\usage{
run_sim(
  n,
  n_length,
  mean_si,
  sd_si,
  R0,
  k = 0.16,
  tf = 37,
  max_potential_cases,
  delay_sample
)
}
\arguments{
\item{n}{Numeric, number of initial cases}

\item{n_length}{Numeric, the number of days in the initial seeding event}

\item{mean_si}{Numeric, the mean of the serial interval}

\item{sd_si}{Numeric, the sd of the serial interval}

\item{R0}{Numeric, the estimated reproduction number}

\item{k}{Numeric, the dispersion of the negative binomial branching process}

\item{tf}{Numeric, the end time of the branching process}

\item{max_potential_cases}{Numeric, the maximum number of cases}

\item{delay_sample}{A function to sample from reporting delays}
}
\value{
A dataframe containing the simulation time and outbreak size
}
\description{
Negative binomial branching process model
}
\examples{

## Code
run_sim
}
\author{
Sebastian Funk, Sam Abbott, James Munday
}
