% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_end_r0.R
\name{summarise_end_r0}
\alias{summarise_end_r0}
\title{Summarise R0 estimates across samples}
\usage{
summarise_end_r0(sims)
}
\arguments{
\item{sims}{A dataframe as produced by \code{scenario_analysis}.}
}
\value{
A dataframe of R0 estimates for each scenario
}
\description{
Summarise R0 estimates across samples
}
\examples{

## Code
summarise_end_r0
}
\author{
Sam Abbott
}
