% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WuhanSeedingVsTransmission-package.R
\docType{package}
\name{WuhanSeedingVsTransmission-package}
\alias{WuhanSeedingVsTransmission}
\alias{WuhanSeedingVsTransmission-package}
\title{WuhanSeedingVsTransmission: The transmissibility of novel Coronavirus in the early stages of the 2019-20 outbreak in Wuhan}
\description{
The transmissibility of novel Coronavirus in the early stages of the 2019-20 outbreak 
      in Wuhan: Exploring initial point-source exposure sizes and durations using scenario analysis
}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

Authors:
\itemize{
  \item Joel Hellewell \email{joel.hellewell@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-2683-0849}{ORCID})
  \item James Munday \email{james.munday@lshtm.ac.uk}
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}

}
\keyword{internal}
